refactor(vehicle): extract enums + enum converters out of vehicle_profile (#3234)
vehicle_profile.dart was a 491-line freezed entity (3 grandfathered file_length
bumps, decomposition forced by #3141) that also carried the standalone enums
and their JSON converters.

- vehicle_enums.dart (132): VehicleType / VehicleCalibrationMode / ConnectorType
  + their enum-only converters (VehicleTypeJsonConverter,
  VehicleCalibrationModeJsonConverter, ConnectorTypeSetConverter).
- vehicle_profile.dart (491 → 377): the freezed VehicleProfile +
  ChargingPreferences + its converter + the auto-record extension + the legacy
  migration only. Re-exports vehicle_enums.dart, so the ~200 sites importing
  the enums via this model resolve them unchanged — zero call-site edits.

Pure relocation: clean codegen produced no .g.dart / .freezed.dart drift, and
the JSON round-trip tests stay green. Removed vehicle_profile's grandfather
entry from the file_length snapshot (both files < 400 cap).

This is the entity half of #3234; the sibling edit_vehicle_screen decomposition
stays tracked there — it needs a form-state-model design (its bulk is intertwined
ConsumerState logic, not a mechanical split), so it's deliberately left for a
focused follow-up rather than forced here.
